Fixing some warnings

+++ Amethyst/Amethyst/src/Amethyst/PerspectiveCamera.cpp
@@ -14,7 +14,7 @@ namespace Amethyst
 		front(0.0f, 0.0f, -1.0f), hFov(glm::radians(70.0f)), vFov(0.0f), nearPlane(0.5f), farPlane(1000.0f), mouseInitialPos(0.0f)
 	{
 		Window& window = Application::Get().GetWindow();
-		float aspectRatio = window.GetWidth() / window.GetHeight();
+		float aspectRatio = (float)window.GetWidth() / (float)window.GetHeight();
 		vFov = 2 * glm::atan(glm::tan(hFov / 2) * aspectRatio);
 		projectionMatrix = glm::perspective(vFov, aspectRatio, nearPlane, farPlane);
 		viewMatrix = ComputeViewMatrix();
